+++ src/main/java/com/tmproject/api/board/controller/BoardController.java
@@ -69,15 +69,6 @@ public class BoardController {
         @AuthenticationPrincipal MemberDetailsImpl memberDetails,
         @PageableDefault(sort = "createdAt") Pageable pageable
     ) {
-        String myInfo = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
-        System.out.println(myInfo);
-        // 로그인을 했는데? 여기는 안들어가?
-        // 여기가 java.lang.NullPointerException: Cannot invoke "com.tmproject.Common.Security.MemberDetailsImpl.getMember()"
-        // because "memberDetails" is null
-        // ? MemberDetails를 못받아 오는 이유가 뭐지?
-        // 로그인한 사용자의 정보를 가져오는 건데...?
-        // memberService에서는 잘 받아온단 말야? 뭔가 이상한데?
-        log.info("memberDetails : "+ memberDetails);
         Member member = memberDetails.getMember();
         List<BoardViewResponseDto> BoardViewResponseDto = boardService.getFollowersBoards(member.getId(), pageable);
