+++ src/main/java/com/tmproject/api/follow/service/FollowService.java
@@ -59,6 +59,4 @@ public class FollowService {
             follower);
         return optionalFollow.isPresent();
     }
-
-
 }
