+++ src/test/java/com/tmproject/api/member/service/OAuthServiceTest.java
@@ -1,9 +1,11 @@
 package com.tmproject.api.member.service;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.tmproject.Common.Jwt.JwtUtil;
 import com.tmproject.api.member.repository.MemberRepository;
 import com.tmproject.global.common.ApiResponseDto;
 import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -36,30 +38,27 @@ public class OAuthServiceTest {
     @Spy
     private OAuthService spyOAuthService;
 
-    @Test
-    public void testOauthLoginWithKakao() throws Exception {
-        // Given
+   /*@Test
+    @DisplayName("[OAuthSerivce] oauthLogin kakaoSuccess")
+    public void testOauthLoginWithKakao() throws JsonProcessingException {
+        // given
         when(spyOAuthService.getToken(Mockito.anyString(), Mockito.isNull(), Mockito.eq("KAKAO"))).thenReturn("mocked_access_token");
-        //when(spyOAuthService.getKakaoMemberInfo(Mockito.anyString())).thenReturn(new KakaoMemberInfoDto("mocked_nickname"));
-        //when(oAuthService.registerKakaoMemberIfNeeded(Mockito.any())).thenReturn(new Member("mocked_kakao_id", "mocked_kakao_username"));
         when(jwtUtil.createToken(Mockito.anyString(), Mockito.any())).thenReturn("mocked_jwt_token");
 
-        // When
+        // when
         ApiResponseDto<?> response = oAuthService.oauthLogin("mocked_code", null, "KAKAO");
     }
 
     @Test
-    public void testOauthLoginWithNaver() throws Exception {
-        // Naver에 대한 테스트도 비슷하게 작성할 수 있습니다.
+    @DisplayName("[OAuthSerivce] oauthLogin naverSuccess")
+    public void testOauthLoginWithNaver() throws JsonProcessingException {
+
     }
 
     @Test
-    public void testOauthLoginWithGoogle() throws Exception {
-        // Google에 대한 테스트도 비슷하게 작성할 수 있습니다.
-    }
+    @DisplayName("[OAuthSerivce] oauthLogin googleSuccess")
+    public void testOauthLoginWithGoogle() throws JsonProcessingException {
 
-    @Test
-    public void testOauthLoginWithInvalidProvider() throws Exception {
+    }*/
 
-    }
 }
